requirements.txt
aiogram
aiohttp
Pillow
torch
torchvision
